+++ tests/Doubles/Meta/ClassInterfaceMetaInvalidScopeRootVO.php
@@ -1,0 +1,15 @@
+<?php declare(strict_types = 1);
+
+namespace Tests\Orisai\ObjectMapper\Doubles\Meta;
+
+use Orisai\ObjectMapper\MappedObject;
+
+final class ClassInterfaceMetaInvalidScopeRootVO implements MappedObject, ClassInterfaceMetaInvalidScopeInterfaceVO
+{
+
+	public function before(): void
+	{
+		// Noop
+	}
+
+}