+++ tests/Doubles/Meta/ClassTraitMetaInvalidScopeVO.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types = 1);
+
+namespace Tests\Orisai\ObjectMapper\Doubles\Meta;
+
+abstract class ClassTraitMetaInvalidScopeVO
+{
+
+	use ClassTraitMetaInvalidScopeTraitVO;
+
+}