+++ tests/Doubles/Meta/FieldMetaInvalidScopeVO.php
@@ -1,0 +1,13 @@
+<?php declare(strict_types = 1);
+
+namespace Tests\Orisai\ObjectMapper\Doubles\Meta;
+
+use Orisai\ObjectMapper\Rules\StringValue;
+
+abstract class FieldMetaInvalidScopeVO
+{
+
+	/** @StringValue() */
+	public string $field;
+
+}