+++ tests/Doubles/Meta/FieldTraitMetaInvalidScopeRootVO.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types = 1);
+
+namespace Tests\Orisai\ObjectMapper\Doubles\Meta;
+
+use Orisai\ObjectMapper\MappedObject;
+
+final class FieldTraitMetaInvalidScopeRootVO extends FieldTraitMetaInvalidScopeVO implements MappedObject
+{
+
+}