+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/orisai/object-mapper/compare/0.1.0...HEAD)
 
+### Added
+
+- Ensure metadata are defined in scope of `MappedObject`
+
 ## [0.1.0](https://github.com/orisai/object-mapper/releases/tag/0.1.0) - 2023-07-07
 
 Initial release